+++ ext/singularity/deployer.go
@@ -90,10 +90,13 @@ func NewDeployer(c rectificationClient, ls logging.LogSink, options ...DeployerO
 func (r *deployer) Rectify(pair *sous.DeployablePair) sous.DiffResolution {
 	postID := ""
 	version := ""
+	var user sous.User
 	if pair.Post != nil {
 		postID = pair.Post.ID().String()
 		version = pair.Post.DeploySpec().Version.String()
+		user = pair.Post.Deployment.User
 	}
+
 	if pair.UUID == uuid.Nil {
 		pair.UUID = uuid.NewV4()
 	}
@@ -110,7 +113,7 @@ func (r *deployer) Rectify(pair *sous.DeployablePair) sous.DiffResolution {
 		messages.ReportLogFieldsMessage("SameKind", logging.InformationLevel, r.log, postID, version, resolution)
 		return resolution
 	case sous.AddedKind:
-		messages.ReportLogFieldsMessageToConsole(fmt.Sprintf("Starting an AddedKind %s:%s", postID, version), logging.ExtraDebug1Level, r.log, pair)
+		messages.ReportLogFieldsMessageToConsole(fmt.Sprintf("Starting an AddedKind %s:%s by user %s", postID, version, user), logging.ExtraDebug1Level, r.log, pair)
 		result := sous.DiffResolution{DeploymentID: pair.ID()}
 		if err := r.RectifySingleCreate(pair); err != nil {
 			result.Desc = "not created"
@@ -131,7 +134,7 @@ func (r *deployer) Rectify(pair *sous.DeployablePair) sous.DiffResolution {
 		reportDiffResolutionMessage("Result of create", result, logging.InformationLevel, r.log)
 		return result
 	case sous.RemovedKind:
-		messages.ReportLogFieldsMessageToConsole(fmt.Sprintf("Starting an RemoveKind %s:%s", postID, version), logging.ExtraDebug1Level, r.log, pair)
+		messages.ReportLogFieldsMessageToConsole(fmt.Sprintf("Starting an RemoveKind %s:%s by user %s", postID, version, user), logging.ExtraDebug1Level, r.log, pair)
 		result := sous.DiffResolution{DeploymentID: pair.ID()}
 		if err := r.RectifySingleDelete(pair); err != nil {
 			result.Error = sous.WrapResolveError(&sous.DeleteError{Deployment: pair.Prior.Deployment.Clone(), Err: err})
@@ -144,7 +147,7 @@ func (r *deployer) Rectify(pair *sous.DeployablePair) sous.DiffResolution {
 		return result
 	case sous.ModifiedKind:
 		result := sous.DiffResolution{DeploymentID: pair.ID()}
-		messages.ReportLogFieldsMessageToConsole(fmt.Sprintf("Starting a ModifiedKind %s:%s", postID, version), logging.ExtraDebug1Level, r.log, pair)
+		messages.ReportLogFieldsMessageToConsole(fmt.Sprintf("Starting a ModifiedKind %s:%s by user %s", postID, version, user), logging.ExtraDebug1Level, r.log, pair)
 		if err := r.RectifySingleModification(pair); err != nil {
 			dp := &sous.DeploymentPair{
 				Prior: pair.Prior.Deployment.Clone(),