+++ test/smoke/testsingularityassertions_test.go
@@ -3,6 +3,7 @@
 package smoke
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/opentable/go-singularity/dtos"
@@ -44,6 +45,14 @@ func assertNilHealthCheckOnLatestDeploy(t *testing.T, f TestFixture, did sous.De
 	gotHealthcheck := dep.Deploy.Healthcheck
 	if gotHealthcheck != nil {
 		t.Fatalf("got Healthcheck = %v; want nil", gotHealthcheck)
+	}
+}
+
+func assertUserOnLatestDeploy(t *testing.T, f TestFixture, did sous.DeploymentID) {
+	t.Helper()
+	dep := f.Singularity.GetLatestDeployForDeployment(t, did)
+	if dep.DeployMarker.User != fmt.Sprintf("sous_%s", f.UserEmail) {
+		t.Errorf("got user %s; want sous_%s", dep.DeployMarker.User, f.UserEmail)
 	}
 }