+++ CHANGELOG.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface
 
-
 ## [Unreleased](//github.com/opentable/sous/compare/0.5.101...HEAD)
+### Added
+* Server: Send user to singularity for Deployment
 ### Fixed
 * Server: the database storage engine wasn't recording the advisory whitelists.